Refactor: LetterCell 복잡한 로직 정리
